* added missing maven dependency * change structure to match Maven standards * added IntelliJ project files * removed compilation folder (target) from git * stopped for now to work on code warnings; will do step by step
