+++ src/main/java/io/github/rowak/nanoleafdesktop/ui/listener/ComponentChangeListener.java
@@ -6,5 +6,5 @@ import javax.swing.event.ChangeListener;
 public interface ComponentChangeListener extends ChangeListener
 {
 	@Override
-	public void stateChanged(ChangeEvent e);
+	void stateChanged(ChangeEvent e);
 }
